copy imperative typedef test
