+++ packages/houdini-core/plugin/documents/typecript/documents.go
@@ -1,0 +1,5 @@
+package typescript
+
+func GenerateDocumentTypeDefs() ([]string, error) {
+	return nil, nil
+}
