+++ packages/houdini-core/plugin/documents/typecript/documents_test.go
@@ -1,0 +1,1 @@
+package typescript_test
